Remove with-lock and with-jruby-instance macros
Remove `with-lock` and `with-jruby-instance` macros from the
jruby puppet service, and instead add aliases for them to the `jruby-utils`
versions of these macros.

Add a `get-pool-context` service function to the jruby puppet service so that
users of the service can get the pool context and pass it to the `with-lock`
and `with-jruby-instance` macros.

Move the puppetserver `jruby-pool-int-test` to `jruby-puppet-pool-int-test` to
not conflict with the test namespace in jruby utils.