(SERVER-1571) Update shell utils test to work independent of environment
Previously, the shell utils `sets-env-correctly` test failed if the
environment in which it was run had additional env vars that the JVM always
added - e.g. SHLVL. This commit updates the test to only test that the var
being set is present, and to ignore other variables dependent on the test run
environment.